Update README.md
Updating the README to reflect a change in policy. The old method allowed features and hotfixes to be merged at the discretion of the programmer. This has now been relegated to the repository administrator so there are now additional steps to take to make sure the process flows through the proper chain of command.